User controller tugadi

+++ src/services/user.service.js
@@ -25,9 +25,30 @@ export const getUserByEmailService=async(email)=>{
     }
 }
 
-export const updateUserService=async()=>{
+export const updateUserService=async({role,first_name,last_name,email,password,phone,date_of_birth,created_at,updated_at})=>{
     try {
-        
+    const res=await findByemail(email)
+        if(res.length>=1){
+            if(role=='admin'){
+                if(res[0].role=='user'){
+                    const result=await connection.select("*").from("customer").where({email}).update({first_name,last_name,email,password,phone,date_of_birth,created_at,updated_at})
+                    return result
+                }else{
+                    return "Siz ushbu userni yangilay olmaysiz"
+                }
+            }else if(role=='manager'){
+                if(res[0].role=='admin' || res[0].role=='user'){
+                    const result=await connection.select("*").from("customer").where({email}).update({first_name,last_name,email,password,phone,date_of_birth,created_at,updated_at})
+                    return result
+                }else{
+                    return "Siz ushbu userni yangilay olmaysiz"
+                }
+            }else{
+                return "Userni Yangilashda xatolik"
+            }
+        }else{
+            "Yangilandigan User topilmadi"
+        }  
     } catch (error) {
         return error.message
     }
@@ -37,22 +58,26 @@ export const updateUserService=async()=>{
 export const deleteUserService=async(email,role)=>{
     try {
         const res=await findByemail(email)
-        if(role=='admin'){
-            if(res[0].role=='user'){
-                const result=await deleteByemail(email)
-                return result
+        if(res.length>=1){
+            if(role=='admin'){
+                if(res[0].role=='user'){
+                    const result=await deleteByemail(email)
+                    return result
+                }else{
+                    return "Siz ushbu userni o'chira olishga huquqingiz yoq!!!"
+                }
+            }else if(role=='manager'){
+                if(res[0].role=='admin' || res[0].role=='user'){
+                    const result=await deleteByemail(email)
+                    return result
+                }else{
+                    return "Siz ushbu userni o'chira olmaysiz"
+                }
             }else{
-                return "Siz ushbu userni o'chira olishga huquqingiz yoq!!!"
-            }
-        }else if(role=='manager'){
-            if(res[0].role=='admin' || res[0].role=='user'){
-                const result=await deleteByemail(email)
-                return result
-            }else{
-                return "Siz ushbu userni o'chira olmaysiz"
+                return "Userni o'chirishda xatolik"
             }
         }else{
-            return "Userni o'chirishda xatolik"
+            "O'chiriladigan User topilmadi"
         }
     } catch (error) {
         return error.message
